requirements.txt
flask>=3.0
python-dotenv>=1.0
google-genai>=1.0
Pillow>=10.0
gunicorn>=21.0
boto3>=1.34
